(limited) updates of files in /tests

+++ tests/conftest.py
@@ -1,36 +1,55 @@
+"""Test configuration and fixtures."""
+
 import logging
 import os
 from pathlib import Path
 from shutil import rmtree
-from typing import Any
 
 import pytest
+import torch
+import torch.cuda
 from crane_controller.crane_factory import build_crane
 
+CUDA_AVAILABLE: bool = torch.cuda.is_available()
 
-@pytest.fixture(scope="package", autouse=True)
-def chdir() -> None:
+TORCH_DEVICES: list[str] = ["cuda", "cpu"] if CUDA_AVAILABLE else ["cpu"]
+
+
+@pytest.fixture(scope="class", params=TORCH_DEVICES)
+def vary_torch_default_device(request: pytest.FixtureRequest):
+    torch.set_default_device(request.param)
+    yield
+    torch.set_default_device("cpu")  # reset to default device after test
+
+
+@pytest.fixture(scope="session", autouse=True)
+def chdir():
     """
     Fixture that changes the current working directory to the 'test_working_directory' folder.
-    This fixture is automatically used for the entire package.
+    This fixture is automatically used for the entire session.
     """
+    original_cwd = Path.cwd()
     os.chdir(Path(__file__).parent.absolute() / "test_working_directory")
+    try:
+        yield
+    finally:
+        os.chdir(original_cwd)  # reset to original working directory after tests
 
 
-@pytest.fixture(scope="package", autouse=True)
+@pytest.fixture(scope="session", autouse=True)
 def test_dir() -> Path:
     """
     Fixture that returns the absolute path of the directory containing the current file.
-    This fixture is automatically used for the entire package.
+    This fixture is automatically used for the entire session.
     """
     return Path(__file__).parent.absolute()
 
 
-output_dirs = [
+output_dirs: list[str] = [
     "results",
     "data",
 ]
-output_files = [
+output_files: list[str] = [
     "*test*.pdf",
 ]
 
@@ -55,7 +74,8 @@ def _remove_output_dirs_and_files() -> None:
     for pattern in output_files:
         for file in Path.cwd().glob(pattern):
             _file = Path(file)
-            _file.unlink(missing_ok=True)
+            if _file.is_file():
+                _file.unlink(missing_ok=True)
 
 
 @pytest.fixture(autouse=True)
@@ -79,10 +99,15 @@ def crane():
     return build_crane
 
 
-def pytest_addoption(parser: Any):
-    parser.addoption("--show", action="store_true", default=False)
+def pytest_addoption(parser: pytest.Parser) -> None:
+    parser.addoption(
+        "--show",
+        action="store_true",
+        default=False,
+        help="Command line switch to show plots during tests, and dump additional results to console. By default, False.",
+    )
 
 
 @pytest.fixture(scope="session")
-def show(request: Any):
+def show(request: pytest.FixtureRequest) -> bool:
     return request.config.getoption("--show")
